@@ -10,23 +10,36 @@
   "author": "Kasey C.",
   "license": "ISC",
   "dependencies": {
-    "alt": "^0.16.5",
     "lodash": "^3.6.0",
-    "react": "^0.13.1"
+    "material-ui": "^0.10.4",
+    "radium": "^0.13.7",
+    "react": "^0.13.1",
+    "react-redux": "^1.0.0",
+    "react-tap-event-plugin": "^0.1.7",
+    "redux": "^1.0.1"
   },
   "devDependencies": {
     "babel-core": "^5.1.10",
     "babel-eslint": "^3.0.1",
     "babel-loader": "^5.0.0",
+    "babel-runtime": "^5.8.20",
     "browser-sync": "^2.6.4",
-    "eslint": "^0.19.0",
-    "eslint-loader": "^0.10.0",
-    "eslint-plugin-react": "^2.1.1",
+    "chai": "^3.0.0",
+    "eslint": "^0.21.0",
+    "eslint-loader": "^0.11.0",
+    "eslint-plugin-react": "^2.3.0",
+    "expect": "^1.6.0",
+    "gulp": "^3.8.11",
+    "gulp-eslint": "^0.12.0",
+    "jsdom": "^3.1.2",
+    "mocha": "^2.2.5",
     "webpack": "^1.8.5"
   },
   "scripts": {
+    "test": "mocha --recursive --compilers js:babel-core/register ./assets/js/__tests__/*.js",
+    "reinstall": "rm -rf ./node_modules && npm install",
     "build": "./node_modules/.bin/webpack",
-    "develop": "./node_modules/.bin/webpack --watch --verbose & npm run browser-sync",
+    "develop": "./node_modules/.bin/webpack --watch --verbose & npm run browser-sync & gulp watch",
     "lint": "./node_modules/.bin/eslint ./assets; true;",
     "browser-sync": "./node_modules/.bin/browser-sync start --config='./bs-config.js'"
   }
