@@ -1,47 +1,85 @@
-let React  = require('react');
+const React  = require('react');
+const Radium = require('Radium');
+const mui    = require('material-ui');
 
-let GarageListActions = require('../actions/GarageListActions');
-let GarageListStore = require('../stores/GarageListStore');
+const ChildContextMixin = require('../mixins/ChildContextMixin');
+
+const List        = mui.List;
+const ListItem    = mui.ListItem;
+const ListDivider = mui.ListDivider;
+const Paper       = mui.Paper;
+const Avatar      = mui.Avatar;
+const TextField   = mui.TextField;
+
+const styles = {
+  searchBar: {
+    marginBottom: '10px'
+  }
+};
 
 /**
  * Main GarageList node, uses Bootstrap's list-group
  */
-let GarageList = React.createClass({
+const GarageList = React.createClass({
 
   propTypes: {
+    searchText: React.PropTypes.string.isRequired,
+    vehicles: React.PropTypes.array.isRequired,
+    searchVehicles: React.PropTypes.func.isRequired
   },
 
-  getInitialState() {
-    return GarageListStore.getState();
-  },
-
-  componentDidMount() {
-    GarageListStore.listen(this.onStoreChange);
-  },
-
-  componentWillUnmount() {
-    GarageListStore.unlisten(this.onStoreChange);
-  },
-
-  onStoreChange(state) {
-    this.setState(state);
-  },
+  mixins: [ChildContextMixin],
 
   render() {
+
+    let {
+      searchText,
+      vehicles,
+      searchVehicles
+    } = this.props;
+
     return (
       <div className="col-sm-6">
-        <GarageList.SearchBar searchText={this.state.searchText} />
-        <div style={{marginTop: 10}}>
-          <div className="list-group">
-            {this.state.vehicles.map((vehicle) =>
-            <a href="#" className="list-group-item">
-              <h4 className="list-group-item-heading">
-                {vehicle.year} {vehicle.make} {vehicle.model}
-              </h4>
-            </a>
-            )}
-          </div>
-        </div>
+        <GarageList.SearchBar
+          searchText={searchText}
+          onSearchVehicles={searchVehicles}
+          />
+
+        <Paper>
+          <List>
+            {vehicles.map((vehicle, index) => {
+
+              const {
+                make,
+                model
+              } = vehicle.links;
+
+              return (
+                <div>
+                  <ListItem
+                    key={`${vehicle.year}_${make.name}_${make.model}`}
+                    primaryText={
+                      <h2>{`${vehicle.year} ${make.name} ${model.name}`}</h2>
+                    }
+                    secondaryText={
+                      <p>
+                        <div><strong>VIN:</strong> {vehicle.vin}</div>
+                        <div><strong>Mileage:</strong> {vehicle.mileage}</div>
+                      </p>
+                    }
+                    leftAvatar={
+                      <Avatar>{make.name[0]}</Avatar>
+                    }
+                  />
+                  {(index < (vehicles.length - 1))
+                    ? <ListDivider />
+                    : null}
+                </div>
+              );
+
+            })}
+          </List>
+        </Paper>
       </div>
     );
   }
@@ -55,24 +93,23 @@ let GarageList = React.createClass({
 GarageList.SearchBar = React.createClass({
 
   propTypes: {
-    searchText: React.PropTypes.string
+    searchText: React.PropTypes.string,
+    onSearchVehicles: React.PropTypes.func.isRequired
   },
 
-  getDefaultProps() {
-    return {
-      searchText: ''
-    };
+  _handleOnChange(searchText) {
+    return this.props.onSearchVehicles(searchText);
   },
 
   render() {
     return (
-      <input
-        type="text"
-        className="form-control"
+      <TextField
+        hintText="Search vehicles..."
         value={this.props.searchText}
-        onChange={(e) => GarageListActions.searchVehicles(e.target.value)}
-        placeholder="Search vehicles..."
-        />
+        onChange={e => this._handleOnChange(e.target.value)}
+        style={styles.searchBar}
+        fullWidth={true}
+      />
     );
   }
 
@@ -89,4 +126,4 @@ GarageList.Vehicle = React.createClass({
 
 });
 
-module.exports = GarageList;
+export default Radium(GarageList);
