@@ -1,17 +1,52 @@
-module.exports = [
-  {
-    year: 1998,
-    make: 'Dodge',
-    model: 'RAM 1500'
-  },
-  {
-    year: 1999,
-    make: 'Buick',
-    model: 'Century'
-  },
-  {
-    year: 2015,
-    make: 'Ford',
-    model: 'Mustang'
+export default {
+  GarageList: {
+    searchText: '',
+    vehicles: [
+      {
+        year:    1998,
+        mileage: 12345,
+        vin:     '1e23rt5c89',
+        links:   {
+          make: {
+            id:   1,
+            name: 'Dodge'
+          },
+          model: {
+            id:   1,
+            name: 'Ram 1500'
+          }
+        }
+      },
+      {
+        year:    1999,
+        mileage: 12345,
+        vin:     '1e23rt5c89',
+        links:   {
+          make: {
+            id:   2,
+            name: 'Buick'
+          },
+          model: {
+            id:   2,
+            name: 'Century'
+          }
+        }
+      },
+      {
+        year:    2015,
+        mileage: 12345,
+        vin:     '1e23rt5c89',
+        links:   {
+          make: {
+            id:   3,
+            name: 'Ford'
+          },
+          model: {
+            id:   3,
+            name: 'Mustang'
+          }
+        }
+      }
+    ]
   }
-];
+};
